Finished generator and size algorithm.
Generator algorithm bugs worked out and appears to be working now. Needs
some work on the comments to make it more clear what it is doing. Perhaps
a visual at the top is in order.

Implemented the size() method via binomial coefficient calculations and
included nice math comment to go with it.